requirements.txt
3==0.0.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2
django-dump-load-utf8==0.0.4
django-filter==22.1
djangorestframework==3.13.1
dnspython==2.2.1
drf-yasg==1.21.3
eventlet==0.33.1
flower==1.2.0
greenlet==1.1.2
gunicorn==20.1.0
humanize==4.3.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycodestyle==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
